fix: ios: lint
