--- ios/NativeSigner/Components/ModalBackdrop.swift
@@ -1,22 +1,0 @@
-//
-//  ModalBackdrop.swift
-//  NativeSigner
-//
-//  Created by Alexander Slesarev on 13.10.2021.
-//
-
-import SwiftUI
-
-struct ModalBackdrop: View {
-    var body: some View {
-        Rectangle().foregroundColor(Color("backgroundColor"))
-    }
-}
-
-/*
-struct ModalBackdrop_Previews: PreviewProvider {
-    static var previews: some View {
-        ModalBackdrop()
-    }
-}
-*/
