+++ ios/NativeSigner/Models/Documents.swift
@@ -5,14 +5,27 @@
 //  Created by Alexander Slesarev on 11.8.2021.
 //
 
+/**
+ * Handle decoding of large hardcoded docs
+ */
+
 import Foundation
 
+/**
+ * Screen model state for documents screen
+ * Since this is hardcoded, we heep it in ios logic at least for now
+ * Moving it to backend will of course have benetif of reducing code reuse
+ * Let's consider it later as now it just works
+ */
 enum ShownDocument {
     case toc
     case pp
     case about
 }
 
+/**
+ * Fetch docs from assets
+ */
 extension SignerDataModel {
     func getTaC() -> AttributedString {
         if let path = Bundle.main.path(forResource: "terms-and-conditions", ofType: "txt") {
