+++ ios/NativeSigner/Models/ModalModels/MNetworkMenu.swift
@@ -11,3 +11,10 @@ struct MNetworkMenu: Decodable, Hashable {
     var networks: [Network]
 }
 
+struct Network: Codable, Hashable {
+    var key: String
+    var logo: String
+    var order: Int
+    var selected: Bool
+    var title: String
+}
