--- ios/NativeSigner/Models/QRView.swift
@@ -1,20 +1,0 @@
-//
-//  QRView.swift
-//  NativeSigner
-//
-//  Created by Alexander Slesarev on 4.8.2021.
-//
-
-import SwiftUI
-
-struct QRView: View {
-    var body: some View {
-        Text("This should be a QR")
-    }
-}
-
-struct QRView_Previews: PreviewProvider {
-    static var previews: some View {
-        QRView()
-    }
-}
